properties created (edited)

+++ database/migrations/2018_10_31_005848_create_cars_table.php
@@ -17,7 +17,7 @@ class CreateCarsTable extends Migration
             $table->increments('id');
             $table->string('make');
             $table->string('model');
-            $table->double('year');
+            $table->bigInteger('year');
             $table->timestamps();
         });
     }
